fix: cast float32 NT tensors to model dtype at module boundaries
The transmitter system (TransmitterSystem.vector()) returns float32.
Three sites multiplied these NT signals against bfloat16 model tensors,
promoting the result to float32 and crashing subsequent bmm/matmul calls:

- workspace.py: ne_temp (float32) * slot_queries (bfloat16) → bmm crash;
  cast candidates and ne_temp to slot_queries.dtype at _forward entry.
- thalamus.py: nt_levels (float32) used in temp/boost arithmetic with
  bfloat16 logits; cast nt_levels to x.dtype before slicing.
- receptors.py: gain() returns float32 when nt_levels is float32;
  cast g to x.dtype in modulate() before the scalar multiply.

+++ neuroslm/neurochem/receptors.py
@@ -233,7 +233,7 @@ class ReceptorBank(nn.Module):
 
     def modulate(self, x: torch.Tensor, nt_levels: torch.Tensor) -> torch.Tensor:
         """Apply gain to last-dim of x."""
-        g = self.gain(nt_levels)
+        g = self.gain(nt_levels).to(dtype=x.dtype)
         shape = [g.size(0)] + [1] * (x.dim() - 1)
         return x * g.view(shape)
